Various Improvements - Version 0.6.5
New bidders added
New textures
Performances issues addressed

+++ Assets/Scripts/Misc/GPGController.cs
@@ -17,7 +17,7 @@ namespace IrishFarmSim
 
 		public static void PostScore(double amount){
 			Debug.Log("Submitting Score");
-			Social.ReportScore((int)amount, "CgkIuaLFq6kIEAIQBg", (bool result) => {
+			Social.ReportScore((long)amount, "CgkIuaLFq6kIEAIQBg", (bool result) => {
 				Debug.Log("Score submitted?: " + result);
 			});
 		}
